implement practice4

+++ src/camera/camera.cpp
@@ -1,13 +1,13 @@
 #include "camera.h"
 
-Ray Camera::cast_in_pixel(glm::vec2 p, glm::vec2 random_offset) {
-    glm::vec3 t{};
+Ray Camera::cast_in_pixel(glm::dvec2 p, glm::dvec2 random_offset) {
+    glm::dvec3 t{};
 
-    t[0] = (2.0 * ((float)p.x + random_offset.x) / (float)width - 1.0) * fov_in_radians.x;
-    t[1] = -(2.0 * ((float)p.y + random_offset.y) / (float)height - 1.0) * fov_in_radians.y;
+    t[0] = (2.0 * (p.x + random_offset.x) / (double)width - 1.0) * fov_in_radians.x;
+    t[1] = -(2.0 * (p.y + random_offset.y) / (double)height - 1.0) * fov_in_radians.y;
     t[2] = 1;
 
-    glm::vec3 direction{};
+    glm::dvec3 direction{};
     direction = direction + t.x * right;
     direction = direction + t.y * up;
     direction = direction + t.z * forward;
