+++ src/objects/intersection.h
@@ -1,8 +1,8 @@
 #pragma once
 
 typedef struct {
-    float distance = std::numeric_limits<float>::max();
-    glm::vec3 normal;
+    double distance = std::numeric_limits<double>::max();
+    glm::dvec3 normal;
     bool inside_flag = false;
     int intersected_object_index = -1;
 } Intersection;
