+++ src/utils/color.h
@@ -5,11 +5,11 @@
 
 #include <cmath>
 
-uint8_t normal_to_rgb(float val);
-glm::vec3 normal_to_rgb(glm::vec3 val);
+uint8_t normal_to_rgb(double val);
+glm::dvec3 normal_to_rgb(glm::dvec3 val);
 
-float rgb_to_normal(int val);
-glm::vec3 rgb_to_normal(glm::vec3 val);
+double rgb_to_normal(int val);
+glm::dvec3 rgb_to_normal(glm::dvec3 val);
 
-glm::vec3 saturate(const glm::vec3 &color);
-glm::vec3 aces_tonemap(const glm::vec3 &x);
+glm::dvec3 saturate(const glm::dvec3 &color);
+glm::dvec3 aces_tonemap(const glm::dvec3 &x);
